process_status_email.rb now retrieves IP address correctly; Devices list shows IP address; remove_old_alerts.rb now removes counter data more than 2 years old too, and devices are removed only if it has neither alerts or counters data

+++ remove_old_alerts.rb
@@ -9,8 +9,12 @@ alert_msg regexp 'Load paper'").each do |a|
   a.delete
 end
 
+Counter.where("created_at < (localtime - interval 2 year)").each do |c|
+  c.delete
+end
+
 Device.all.each do |d|
-  if d.alerts.empty? and d.name != 'default'
+  if d.alerts.empty? and d.counters.empty? and d.name != 'default'
     d.destroy
   end
 end
